% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{loom_to_sce}
\alias{loom_to_sce}
\alias{h5ad_to_sce}
\alias{to_sparse_matrix}
\alias{.to_sparse_matrix.dgCMatrix}
\alias{.to_sparse_matrix.dgRMatrix}
\alias{.to_sparse_matrix.DelayedMatrix}
\alias{.to_sparse_matrix.matrix}
\alias{sparse_mtx_to_assay_tbl}
\alias{.sparse_matrix_to_assay_tbl.dgCMatrix}
\alias{sce_rowdata_to_gene_tbl}
\alias{sce_coldata_to_cell_tbl}
\title{Helper functions for the processing of .loom and .h5ad files into
relational database tables}
\usage{
loom_to_sce(loom_filepath)

h5ad_to_sce(h5ad_filepath)

to_sparse_matrix(matrix_obj)

.to_sparse_matrix.dgCMatrix(c_matrix)

.to_sparse_matrix.dgRMatrix(r_matrix)

.to_sparse_matrix.DelayedMatrix(d_matrix)

.to_sparse_matrix.matrix(m_matrix)

sparse_mtx_to_assay_tbl(sparse_matrix)

.sparse_matrix_to_assay_tbl.dgCMatrix(sparse_matrix)

sce_rowdata_to_gene_tbl(sce)

sce_coldata_to_cell_tbl(sce)
}
\arguments{
\item{loom_filepath}{character() absolute path to downloaded .loom file}

\item{h5ad_filepath}{character() absolute path to downloaded .h5ad file}

\item{sparse_matrix}{dgCMatrix sparse matrix representation}

\item{sce}{SingleCellExperiment or SingleCellLoomExperiment object}

\item{sce_obj}{SummarizedExperiment (or subclass) object}

\item{dgCMatrix}{sparse matrix}

\item{dgRMatrix}{sparse matrix}

\item{DelayedMatrix}{sparse matrix}

\item{matrix}{sparse matrix}
}
\value{
SingleCellLoomExperiment object

SingleCellExperiment object

dgCMatrix sparse matrix representation

dgCMatrix sparse matrix representation

dgCMatrix sparse matrix representation

dgCMatrix sparse matrix representation

tibble with row_index, col_index, and value of each non-zero entry
in the assay matrix

tibble of gene annotations

tibble of cell annotations
}
\description{
convert a SummarizedExperiment assay to a
sparse matrix representation

function to change SingleCellLoomExperiment object's assay
matrix into sparse representation (\code{dgCMatrix})

function to change SingleCellLoomExperiment object's assay
matrix into sparse representation (\code{dgRMatrix})

function to change SingleCellLoomExperiment object's assay
matrix into sparse representation (\code{DelayedMatrix})

function to change SingleCellLoomExperiment object's assay
matrix into sparse representation (\code{matrix})

convert a sparse matrix representation of the assay into a
reformatted tibble

convert a dgCMatrix sparse matrix representation of the assay
into a reformatted tibble

function to change SingleCellExperiment or
SingleCellLoomExperiment object's \code{rowData} into a gene annotation table

function to change SingleCellExperiment or
SingleCellLoomExperiment object's \code{colData} into a cell annotation table
}
