% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{files_to_db}
\alias{files_to_db}
\alias{.single_file_to_db}
\title{Populating database with data output of single cell experiments}
\usage{
files_to_db(file_tbl = NULL)

.single_file_to_db(file_path, fileId, projectTitle, projectId)
}
\arguments{
\item{file_tbl}{tbl_hca tibble of files as returned by hca::files()}

\item{file_path}{character() location of experiment output file}

\item{fileId}{character() unique identifier of file}

\item{projectTitle}{character() title of file's associated project}

\item{projectId}{character() unique identifier of the file's associated project}
}
\description{
Populating database with data output of single cell experiments
}
