% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{available_tables}
\alias{available_tables}
\alias{table_description}
\title{Functions for querying and interacting with the database}
\usage{
available_tables()

table_description(table_name)
}
\arguments{
\item{table_name}{character() name of database table to be described}
}
\description{
Functions for querying and interacting with the database
}
\examples{
\dontrun{available_tables()}

\dontrun{table_description("genes_tbl")}

}
