% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sce.R
\name{loom_to_sce}
\alias{loom_to_sce}
\alias{h5ad_to_sce}
\alias{loom_sparse_matrix}
\alias{h5ad_sparse_matrix}
\alias{loom_sparse_mtx_to_assay_tbl}
\alias{h5ad_sparse_mtx_to_assay_tbl}
\alias{sce_rowdata_to_gene_tbl}
\alias{sce_coldata_to_cell_tbl}
\title{Helper functions for the processing of .loom and .h5ad files into
relational database tables}
\usage{
loom_to_sce(loom_filepath)

h5ad_to_sce(h5ad_filepath)

loom_sparse_matrix(loom_filepath)

h5ad_sparse_matrix(h5ad_filepath)

loom_sparse_mtx_to_assay_tbl(sparse_matrix)

h5ad_sparse_mtx_to_assay_tbl(sparse_matrix)

sce_rowdata_to_gene_tbl(sce)

sce_coldata_to_cell_tbl(sce)
}
\arguments{
\item{loom_filepath}{character() absolute path to downloaded .loom file}

\item{h5ad_filepath}{character() absolute path to downloaded .h5ad file}

\item{sparse_matrix}{dgRMatrix sparse matrix representation}

\item{sce}{SingleCellExperiment or SingleCellLoomExperiment object}
}
\value{
SingleCellLoomExperiment object

SingleCellExperiment object

dgCMatrix sparse matrix representation

dgRMatrix sparse matrix representation

tibble with row_index, col_index, and value of each non-zero entry
in the assay matrix

tibble with row_index, col_index, and value of each non-zero entry
in the assay matrix

tibble of gene annotations

tibble of cell annotations
}
\description{
function to change SingleCellLoomExperiment object's assa
matrix into sparse representation (\code{dgCMatrix})

function to change extract the existing sparse matrix
representation (\code{dgRMatrix}) of the SingleCellExperiment object's assay

convert a sparse matrix representation of the assay into a
reformatted tibble

convert a sparse matrix representation of the assay into a
reformatted tibble

function to change SingleCellExperiment or
SingleCellLoomExperiment object's \code{rowData} into a gene annotation table

function to change SingleCellExperiment or
SingleCellLoomExperiment object's \code{colData} into a cell annotation table
}
